Don't use Netty's ChannelLocal for ConnectionContexts
Summary:

ChannelLocal has two modes. In one we were using, the value for a
Channel is cleared when the Channel is closed. This is a problem,
because if a client drops a connection after sending a full request but
before NiftyDispatcher processes a request (e.g. if the Executor given
to NiftyDispatcher takes a long time to begin the task and the client
times out), the ConnectionContext will have been discarded. The
ChanneLocal will make a new ConnectionContext when NiftyProcessor tries
to access it, and that new context will be missing information like the
remote address. This, in turn, leads to the infamous "impossible
exception".

The alternative mode for ChannelLocal is to keep values around until the
Channels get GC'd (using a WeakHashMap). Using this would solve the
problem above, but still has the confusing behavior with object
lifetimes (a ConnectionContext comes into existence magically for any
Channel and may or may not be fully initialized at time of use.)

So, instead of ChannelLocal, use the ChannelHandlerContext Netty
provides for our ConnectionContextHandler. When the socket is first
accepted, ConnectionContextHandler creates a ConnectionContext,
initializes it, and puts it in its ConnectionContextHandler (which is
unique to the pipeline and therefore to the Channel). When other code
wants the context, `ConnectionContexts.get()` can ensure it's been set
before returning. (This won't break compatibility with anything, since
ConnectionContextHandler is first in the pipeline which means the
context will always be set before anything else gets the Channel.)

Test Plan: tests pass

Reviewers: andrewcox, alandau, haijunz

CC:

Task ID: #5426461

Blame Rev: