RFC: Some improvements and refactoring in NiftyDispatcher
The major refactor is splitting `processRequest` into two distinct
methods, `processRequestAndSendResponse` which handles the channel-level
logic and `computeResponseForRequest` which is a (kinda) pure function
from `ThriftMessage` request to `ListenableFuture<ThriftMessage>`.

The smaller refactors are:

- some minor code (e.g. constructing the TProtocols was moved around)

- slightly simplify the "start the timer" code (merge two if blocks)

- change `sendTApplicationException` into `makeTApplicationExceptionResponseInto`

- Instead of using an AtomicBoolean to make sure the response is sent
  only once, delegate that to SettableFuture which can only be set once.

- Catch all the exceptions, and document what needs to catch what.

- Remove isOrderedResponseRequired from writeResponse, since we have the
  ctx

Some minor code (e.g. constructing the TProtocols) was moved around.

This gives us a couple advantages. The primary one is that, wit the
simpler code structure, we can reason far more clearly about the
exception behavior, and we now guarantee that we eventually handle the
request (either send a response or close the socket, but never leave it
dangling.)

Test Plan: not yet tested (at all)

Reviewers: alandau, andrewcox, haijunz