Fix read-blocking in server dispatcher
The intent was to block reads on channels where the client required ordered responses, to prevent the server caching an unlimited number of responses while waiting for some response that needs to be sent before the others. But the code was actually sometimes blocking reads on out-of-order-aware channels, which was a problem because the unblocking code was only operating for order-required channels.