Add ConnectionContext.hasClientDisconnected method
Summary: Pretty much what it sounds like. It's implemented as a boolean in
NiftyConnectionContext, and managed by ConnectionContextHandler (like
NiftyConnectionContext.remoteAddress, but from a channelDisconnected event.) It
would be perhaps slightly nicer at some point to have NCC just have a reference
to the Channel directly, but I wanted to avoid larger refactors here.

Test plan: Code should be pretty straightforward. I read through a bit of the
netty nio stuff, and it looks like channelDisconnected is the right thing to
listen for (it triggers from the same flow for all connected channels, but
somewhat earlier.) Some tests included (which are probably harder to review
than the code, sorry!) This would break any code which has other
implementations of ConnectionContext; I don't see much of a way around that
unless we never want to add anything to CC again.

Reviewers: alandau, andrewcox, hainjuz, tgeorgiou